+++ backend/test/src/test_random_route.cpp
@@ -85,10 +85,10 @@ TEST(test_person, random_test_prioritized_planner) {
         std::vector<Border> border;
         int border_size = static_cast<int>(distribution(generator));
         for (int j = 0; j < border_size; ++j) {
-            int same_coord = distribution(generator);
-            int diff_coord_2 = distribution(generator);
-            int diff_coord_1 = distribution(generator);
-            int direction = wall_direction_distribution(generator);
+            int same_coord = static_cast<int>(distribution(generator));
+            int diff_coord_2 = static_cast<int>(distribution(generator));
+            int diff_coord_1 = static_cast<int>(distribution(generator));
+            int direction = static_cast<int>(wall_direction_distribution(generator));
             if (direction == 0) {
                 border.push_back(Border(
                     Point(same_coord, diff_coord_1),
